synthie::params

Enum AudioEvent

pub enum AudioEvent {
    Panic,
    NoteOn(MidiNote),
    NoteOff(MidiNote),
    LoadPatch(Box<SynthParams>),
    Drum(DrumHit),
    NoteOnChannel(ChannelNo, MidiNote),
    NoteOffChannel(ChannelNo, MidiNote),
    LoadPatchChannel(ChannelNo, Box<SynthParams>),
    ArpSetNotes(ChannelNo, [MidiNote; 4], u8),
    ArpEnabled(ChannelNo, bool),
}

Messages sent from the UI thread to the audio thread over the event channel.

Variants

Panic

Immediately silence all voices and clear active note routing on all channels.

NoteOn(MidiNote)

Begin sustaining a note at the given MIDI note number on channel 0.

NoteOff(MidiNote)

Release the note at the given MIDI note number on channel 0.

LoadPatch(Box<SynthParams>)

Replace the parameter set for channel 0 with a new snapshot.

Drum(DrumHit)

Trigger a one-shot synthesized drum hit.

NoteOnChannel(ChannelNo, MidiNote)

Begin sustaining a note on the given channel at the given MIDI note number.

NoteOffChannel(ChannelNo, MidiNote)

Release the note on the given channel at the given MIDI note number.

LoadPatchChannel(ChannelNo, Box<SynthParams>)

Replace the parameter set for the given channel with a new snapshot.

ArpSetNotes(ChannelNo, [MidiNote; 4], u8)

Load an explicit note list into the arpeggiator on the given channel. Only entries 0..count of notes are used; count must be <= 4.

ArpEnabled(ChannelNo, bool)

Enable or disable the arpeggiator on the given channel at runtime.